spam-filter/requirements.txt
psycopg2==2.9.3
matplotlib~=3.5.2
scikit-learn~=1.0.2
Flask~=2.1.3
python-dotenv~=0.20.0
joblib~=1.1.0
SQLAlchemy~=1.4.39
nltk~=3.7
scipy~=1.8.1
numpy~=1.23.1
textblob~=0.17.1
kafka-python~=2.0.2
Flask-SQLAlchemy~=2.5.1
pandas~=1.4.3
statsmodels~=0.13.2
requests~=2.28.1